Version 3.1.0 ## Update for Versions The plugin can now be run from 1.9 to latest (currently: 1.21.1) We might incounter some issues, if you find any please report them to us!
## Chances and Raritys
All rewards have now chances and raritys so player need to have luck to recive items.
The raritys can be configured in the config.

## Auto Downloader
The plugin will now download all required plugins itself and automaticly update them if a newer version that is older than 3 day comes out.
It will also update itself if a new update comes out.
Old plugins will be moved to a new folder OLD_PLUGINS

## Menus and Messages
We translated all menus and fixed missing messages.
They can now be changed in the menu and messages file.

## Command changes
The command /eggHunt placeEggs was changed to /eggHunt place.
All Permissions got changed as well.
/eggHunt is now the default command it can be changed in the config to your custom commands
This only works if the server reloads or restarts not on plugin reload!!

## Requirements
Added dates to requirements.
Added support for the Requirements Order -> Can now select between OR or AND

## Bug fixed
Fixed a line that was double in the config.yml file
Fixed that firework did damage to player or entitys
Fixed that when a collection is disabled it nevertheless shows the actionbar title and particle

## Name change
The plugin is now called AdvancedHunt because of a new feature
You can now change the current placeholder 'egg' or 'eggs' to what ever you want
For example for christmas like 'present' or 'presents'
This feature is not 100% compleate so feel free to suggest stuff or report stuff

## Other
Added a permission to finding eggs
Added a back button when going back from the collection menu
Added a collection to select in the /eggs reset [player] [collection] command
The eggs in the menu progress and list are now synchronized with the placed eggs
You now dont take damage or loose hunger in place mode

Co-authored-by: TheRedEnd2000 <83455260+theredend2000@users.noreply.github.com>